server/requirements.txt
diffusers==0.28.2
transformers==4.41.1
--extra-index-url https://download.pytorch.org/whl/cu121;
torch==2.2.2
fastapi==0.111.0
uvicorn[standard]==0.30.0
Pillow==10.3.0
accelerate==0.30.1
compel==2.0.2
controlnet-aux==0.0.9
peft==0.11.1
xformers; sys_platform != 'darwin' or platform_machine != 'arm64'
markdown2
safetensors
stable_fast @ https://github.com/chengzeyi/stable-fast/releases/download/v1.0.5/stable_fast-1.0.5+torch222cu121-cp310-cp310-manylinux2014_x86_64.whl ; sys_platform != 'darwin' or platform_machine != 'arm64'
oneflow @ https://github.com/siliconflow/oneflow_releases/releases/download/community_cu121/oneflow-0.9.1.dev20240515+cu121-cp310-cp310-manylinux_2_17_x86_64.manylinux2014_x86_64.whl ; sys_platform != 'darwin' or platform_machine != 'arm64'
onediff @ git+https://github.com/siliconflow/onediff.git@main#egg=onediff ; sys_platform != 'darwin' or platform_machine != 'arm64'
setuptools
mpmath==1.3.0
numpy==1.*
controlnet-aux
ndi-python
syphon-python